b87

+++ src/alg.cpp
@@ -19,7 +19,7 @@ uint64_t fact(uint16_t n) {
 }
 
 double calcItem(double x, uint16_t n) {
-  return pown(x, n) / fact(n);
+  return pown(x, n) / static_cast<double>(fact(n));
 }
 
 double expn(double x, uint16_t count) {
@@ -34,7 +34,7 @@ double sinn(double x, uint16_t count) {
   double sum = 0.0;
   for (uint16_t n = 0; n < count; ++n) {
     int sign = (n % 2 == 0) ? 1 : -1;
-    sum += sign * pown(x, 2 * n + 1) / fact(2 * n + 1);
+    sum += sign * pown(x, 2 * n + 1) / static_cast<double>(fact(2 * n + 1));
   }
   return sum;
 }
@@ -43,7 +43,7 @@ double cosn(double x, uint16_t count) {
   double sum = 0.0;
   for (uint16_t n = 0; n < count; ++n) {
     int sign = (n % 2 == 0) ? 1 : -1;
-    sum += sign * pown(x, 2 * n) / fact(2 * n);
+    sum += sign * pown(x, 2 * n) / static_cast<double>(fact(2 * n));
   }
   return sum;
 }
